updated logic for the graphview
