+++ src/tracksdata/graph/_sql_graph.py
@@ -1563,9 +1563,8 @@ class SQLGraph(BaseGraph):
                     .filter(self.Edge.source_id == source_id, self.Edge.target_id == target_id)
                     .delete()
                 )
-            if deleted == 0:
-                session.rollback()
                 if edge_id is not None:
                     raise ValueError(f"Edge {edge_id} does not exist in the graph")
+            if not deleted:
                 raise ValueError(f"Edge {source_id}->{target_id} does not exist in the graph")
             session.commit()
